Qu'est-ce que c'est

+++ TPEPOD/src/main/java/QueryOneMapper.java
@@ -12,7 +12,7 @@ public class QueryOneMapper implements Mapper<String, Movie, String, Long> {
     public void map(String s, Movie movie, Context<String, Long> context) {
 
         for (String actor : movie.getActors())    {
-            context.emit(actor, (long) movie.getImdbVotes());
+            context.emit(actor, movie.getImdbVotes());
         }
     }
 }
